HealthHanging is done
now health is hanging upon hit. If hit again, while health is hanging - the player takes twice the damage. If the timer of Hanging health runs out - the base damage applies to health. if the player hits an enemy while the health is hanging - he deals +1 damage point to the base damage